third_party/blink/web_tests/clipboard/async-clipboard/async-navigator-clipboard-basics.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Async Clipboard input type validation tests</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../http/tests/resources/permissions-helper.js"></script>
<script>
// Permissions are required in order to invoke navigator.clipboard functions in
// an automated test.
async function getPermissions() {
  await PermissionsHelper.setPermission('clipboard-read', 'granted');
  await PermissionsHelper.setPermission('clipboard-write', 'granted');
}

test(() => {
  assert_not_equals(navigator.clipboard, undefined);
  assert_true(navigator.clipboard instanceof Clipboard);
  assert_equals(navigator.clipboard, navigator.clipboard);
}, 'navigator.clipboard exists');

promise_test(async () => {
  await getPermissions();
  const blob = new Blob(['hello'], {type: 'text/plain'});
  await navigator.clipboard.write({'text/plain': blob});
}, 'navigator.clipboard.write({string : text/plain Blob}) succeeds');

promise_test(async t => {
  await getPermissions();
  await promise_rejects(t, new TypeError(),
                         navigator.clipboard.write());
}, 'navigator.clipboard.write() fails (expect {string : Blob})');

promise_test(async t => {
  await getPermissions();
  await promise_rejects(t, new TypeError(),
                         navigator.clipboard.write(null));
}, 'navigator.clipboard.write(null) fails (expect {string : Blob})');

promise_test(async t => {
  await getPermissions();
  await promise_rejects(t, new TypeError(),
                         navigator.clipboard.write('Bad string'));
}, 'navigator.clipboard.write(DOMString) fails (expect {string : Blob})');

promise_test(async t => {
  await getPermissions();
  const blob = new Blob(['hello'], {type: 'text/plain'});
  await promise_rejects(t, 'NotAllowedError',
                         navigator.clipboard.write(blob));
}, 'navigator.clipboard.write(Blob) fails (expect {string : Blob})');

promise_test(async () => {
  await getPermissions();
  await navigator.clipboard.writeText('New clipboard text');
}, 'navigator.clipboard.writeText(DOMString) succeeds');

promise_test(async t => {
  await getPermissions();
  await promise_rejects(t, new TypeError(),
                         navigator.clipboard.writeText());
}, 'navigator.clipboard.writeText() fails (expect DOMString)');

promise_test(async () => {
  await getPermissions();
  const result = await navigator.clipboard.read();
  assert_true(result instanceof Object);
}, 'navigator.clipboard.read() succeeds');

promise_test(async () => {
  await getPermissions();
  const result = await navigator.clipboard.readText();
  assert_equals(typeof result, 'string');
}, 'navigator.clipboard.readText() succeeds');

</script>
